Assignment_5/rag-windturbine-project/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
faiss-cpu
sentence-transformers
numpy
pandas
PyPDF2
python-docx
aiofiles
pydantic
requests
transformers